+++ lib/config/presets/forgejo/index.ts
@@ -1,6 +1,6 @@
 import { logger } from '../../../logger/index.ts';
 import { getRepoContents } from '../../../modules/platform/forgejo/forgejo-helper.ts';
-import type { RepoContents } from '../../../modules/platform/forgejo/schema.ts';
+import type { RepoContents } from '../../../modules/platform/gitea-common/schema.ts';
 import { ExternalHostError } from '../../../types/errors/external-host-error.ts';
 import type { Nullish } from '../../../types/index.ts';
 import type { Preset, PresetConfig } from '../types.ts';
