+++ lib/modules/platform/gitea-common/helper.spec.ts
@@ -1,6 +1,6 @@
 import * as httpMock from '~test/http-mock.ts';
 import { fakeSha, logger, partial } from '~test/util.ts';
-import { setBaseUrl } from '../../../util/http/forgejo.ts';
+import { GiteaHttp, setBaseUrl } from '../../../util/http/gitea.ts';
 import { toBase64 } from '../../../util/string.ts';
 import {
   closeIssue,
@@ -33,7 +33,7 @@ import {
   updateIssue,
   updateIssueLabels,
   updatePR,
-} from './forgejo-helper.ts';
+} from './helper.ts';
 import type {
   Branch,
   Comment,
@@ -48,11 +48,13 @@ import type {
   User,
 } from './schema.ts';
 
-describe('modules/platform/forgejo/forgejo-helper', () => {
-  const forgejoApiHost = 'https://forgejo.renovatebot.com/';
-  const baseUrl = `${forgejoApiHost}api/v1`;
+describe('modules/platform/gitea-common/helper', () => {
+  const giteaHttp = new GiteaHttp();
 
-  const mockCommitHash = fakeSha('forgejo-helper');
+  const apiHost = 'https://gitea.renovatebot.com/';
+  const baseUrl = `${apiHost}api/v1`;
+
+  const mockCommitHash = fakeSha('gitea-helper');
 
   const mockUser: User = {
     id: 1,
@@ -75,8 +77,8 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
     allow_rebase_explicit: true,
     allow_merge_commits: true,
     allow_squash_merge: true,
-    clone_url: 'https://forgejo.renovatebot.com/some/repo.git',
-    ssh_url: 'git@forgejo.renovatebot.com/some/repo.git',
+    clone_url: 'https://gitea.renovatebot.com/some/repo.git',
+    ssh_url: 'git@gitea.renovatebot.com/some/repo.git',
     default_branch: 'master',
     full_name: 'some/repo',
     archived: false,
@@ -96,7 +98,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
   const otherMockRepo: Repo = {
     ...mockRepo,
     full_name: 'other/repo',
-    clone_url: 'https://forgejo.renovatebot.com/other/repo.git',
+    clone_url: 'https://gitea.renovatebot.com/other/repo.git',
   };
 
   const mockLabel: Label = {
@@ -118,7 +120,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
     title: 'Some PR',
     body: 'Lorem ipsum dolor sit amet',
     mergeable: true,
-    diff_url: `https://forgejo.renovatebot.com/${mockRepo.full_name}/pulls/13.diff`,
+    diff_url: `https://gitea.renovatebot.com/${mockRepo.full_name}/pulls/13.diff`,
     base: { ref: mockRepo.default_branch },
     head: {
       label: 'pull-req-13',
@@ -149,7 +151,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
     status: 'success',
     context: 'some-context',
     description: 'some-description',
-    target_url: 'https://forgejo.renovatebot.com/commit-status',
+    target_url: 'https://gitea.renovatebot.com/commit-status',
     created_at: '2020-03-25T00:00:00Z',
   };
 
@@ -188,14 +190,14 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
   };
 
   beforeEach(() => {
-    setBaseUrl(forgejoApiHost);
+    setBaseUrl(apiHost);
   });
 
   describe('getCurrentUser', () => {
     it('should call /api/v1/user endpoint', async () => {
       httpMock.scope(baseUrl).get('/user').reply(200, mockUser);
 
-      const res = await getCurrentUser();
+      const res = await getCurrentUser(giteaHttp);
       expect(res).toEqual(mockUser);
     });
   });
@@ -205,7 +207,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
       const version = '1.13.01.14.0+dev-754-g5d2b7ba63';
       httpMock.scope(baseUrl).get('/version').reply(200, { version });
 
-      const res = await getVersion();
+      const res = await getVersion(giteaHttp);
 
       expect(res).toEqual(version);
     });
@@ -221,11 +223,15 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(404)
         .get(`/orgs/error`)
         .reply(503);
-      await expect(isOrg(mockRepo.owner.login)).resolves.toEqual(true);
+      await expect(
+        isOrg(giteaHttp, 'gitea', mockRepo.owner.login),
+      ).resolves.toEqual(true);
       // uses cached result
-      await expect(isOrg(mockRepo.owner.login)).resolves.toEqual(true);
-      await expect(isOrg('user')).resolves.toEqual(false);
-      await expect(isOrg('error')).rejects.toThrow(
+      await expect(
+        isOrg(giteaHttp, 'gitea', mockRepo.owner.login),
+      ).resolves.toEqual(true);
+      await expect(isOrg(giteaHttp, 'gitea', 'user')).resolves.toEqual(false);
+      await expect(isOrg(giteaHttp, 'gitea', 'error')).rejects.toThrow(
         'Request failed with status code 503 (Service Unavailable)',
       );
     });
@@ -241,7 +247,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
           data: [mockRepo, otherMockRepo],
         });
 
-      const res = await searchRepos({});
+      const res = await searchRepos(giteaHttp, {});
       expect(res).toEqual([mockRepo, otherMockRepo]);
     });
 
@@ -254,7 +260,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
           data: [otherMockRepo],
         });
 
-      const res = await searchRepos({
+      const res = await searchRepos(giteaHttp, {
         uid: 13,
         archived: false,
       });
@@ -267,7 +273,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         data: [],
       });
 
-      await expect(searchRepos({})).rejects.toThrow(
+      await expect(searchRepos(giteaHttp, {})).rejects.toThrow(
         'Unable to search for repositories, ok flag has not been set',
       );
     });
@@ -277,7 +283,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
     it('should call /api/v1/orgs/[organization]/repos endpoint', async () => {
       httpMock.scope(baseUrl).get('/orgs/some/repos').reply(200, [mockRepo]);
 
-      const res = await orgListRepos('some');
+      const res = await orgListRepos(giteaHttp, 'some');
       expect(res).toEqual([mockRepo]);
     });
   });
@@ -289,7 +295,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}`)
         .reply(200, mockRepo);
 
-      const res = await getRepo(mockRepo.full_name);
+      const res = await getRepo(giteaHttp, mockRepo.full_name);
       expect(res).toEqual(mockRepo);
     });
   });
@@ -303,7 +309,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/contents/${mockContents.path}`)
         .reply(200, { ...mockContents, contentString: undefined });
 
-      const res = await getRepoContents(mockRepo.full_name, mockContents.path);
+      const res = await getRepoContents(
+        giteaHttp,
+        mockRepo.full_name,
+        mockContents.path,
+      );
       expect(res).toEqual(mockContents);
     });
 
@@ -316,6 +326,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, { ...mockContents, contentString: undefined });
 
       const res = await getRepoContents(
+        giteaHttp,
         mockRepo.full_name,
         mockContents.path,
         mockCommitHash,
@@ -332,6 +343,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, otherMockContents);
 
       const res = await getRepoContents(
+        giteaHttp,
         mockRepo.full_name,
         otherMockContents.path,
       );
@@ -346,7 +358,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .post(`/repos/${mockRepo.full_name}/pulls`)
         .reply(200, mockPR);
 
-      const res = await createPR(mockRepo.full_name, {
+      const res = await createPR(giteaHttp, mockRepo.full_name, {
         state: mockPR.state,
         title: mockPR.title,
         body: mockPR.body,
@@ -373,7 +385,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .patch(`/repos/${mockRepo.full_name}/pulls/${mockPR.number}`)
         .reply(200, updatedMockPR);
 
-      const res = await updatePR(mockRepo.full_name, mockPR.number, {
+      const res = await updatePR(giteaHttp, mockRepo.full_name, mockPR.number, {
         state: 'closed',
         title: 'new-title',
         body: 'new-body',
@@ -391,7 +403,9 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .patch(`/repos/${mockRepo.full_name}/pulls/${mockPR.number}`)
         .reply(200, mockPR);
 
-      await expect(closePR(mockRepo.full_name, mockPR.number)).toResolve();
+      await expect(
+        closePR(giteaHttp, mockRepo.full_name, mockPR.number),
+      ).toResolve();
     });
   });
 
@@ -403,7 +417,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200);
 
       await expect(
-        mergePR(mockRepo.full_name, mockPR.number, {
+        mergePR(giteaHttp, mockRepo.full_name, mockPR.number, {
           Do: 'rebase',
         }),
       ).toResolve();
@@ -417,7 +431,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/pulls/${mockPR.number}`)
         .reply(200, mockPR);
 
-      const res = await getPR(mockRepo.full_name, mockPR.number);
+      const res = await getPR(giteaHttp, mockRepo.full_name, mockPR.number);
       expect(res).toEqual(mockPR);
     });
   });
@@ -432,6 +446,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, mockPR);
 
       const res = await getPRByBranch(
+        giteaHttp,
         mockRepo.full_name,
         mockPR.base!.ref,
         mockPR.head!.label,
@@ -448,6 +463,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(404);
 
       const res = await getPRByBranch(
+        giteaHttp,
         mockRepo.full_name,
         mockPR.base!.ref,
         mockPR.head!.label,
@@ -464,6 +480,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(410);
 
       const res = await getPRByBranch(
+        giteaHttp,
         mockRepo.full_name,
         mockPR.base!.ref,
         mockPR.head!.label,
@@ -489,7 +506,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200);
 
       await expect(
-        requestPrReviewers(mockRepo.full_name, mockPR.number, {}),
+        requestPrReviewers(giteaHttp, mockRepo.full_name, mockPR.number, {}),
       ).toResolve();
     });
   });
@@ -501,7 +518,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .post(`/repos/${mockRepo.full_name}/issues`)
         .reply(200, mockIssue);
 
-      const res = await createIssue(mockRepo.full_name, {
+      const res = await createIssue(giteaHttp, mockRepo.full_name, {
         state: mockIssue.state,
         title: mockIssue.title,
         body: mockIssue.body,
@@ -526,12 +543,17 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .patch(`/repos/${mockRepo.full_name}/issues/${mockIssue.number}`)
         .reply(200, updatedMockIssue);
 
-      const res = await updateIssue(mockRepo.full_name, mockIssue.number, {
-        state: 'closed',
-        title: 'new-title',
-        body: 'new-body',
-        assignees: [otherMockUser.login],
-      });
+      const res = await updateIssue(
+        giteaHttp,
+        mockRepo.full_name,
+        mockIssue.number,
+        {
+          state: 'closed',
+          title: 'new-title',
+          body: 'new-body',
+          assignees: [otherMockUser.login],
+        },
+      );
       expect(res).toEqual(updatedMockIssue);
     });
   });
@@ -549,6 +571,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, updatedMockLabels);
 
       const res = await updateIssueLabels(
+        giteaHttp,
         mockRepo.full_name,
         mockIssue.number,
         {
@@ -566,7 +589,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .patch(`/repos/${mockRepo.full_name}/issues/${mockIssue.number}`)
         .reply(200, mockIssue);
 
-      const res = await closeIssue(mockRepo.full_name, mockIssue.number);
+      const res = await closeIssue(
+        giteaHttp,
+        mockRepo.full_name,
+        mockIssue.number,
+      );
       expect(res).toBeUndefined();
     });
   });
@@ -578,7 +605,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/issues?type=issues`)
         .reply(200, [mockIssue]);
 
-      const res = await searchIssues(mockRepo.full_name, {});
+      const res = await searchIssues(giteaHttp, mockRepo.full_name, {});
       expect(res).toEqual([mockIssue]);
     });
 
@@ -588,7 +615,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/issues?state=open&type=issues`)
         .reply(200, [mockIssue]);
 
-      const res = await searchIssues(mockRepo.full_name, {
+      const res = await searchIssues(giteaHttp, mockRepo.full_name, {
         state: 'open',
       });
       expect(res).toEqual([mockIssue]);
@@ -602,7 +629,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/issues/${mockIssue.number}`)
         .reply(200, mockIssue);
 
-      const res = await getIssue(mockRepo.full_name, mockIssue.number);
+      const res = await getIssue(
+        giteaHttp,
+        mockRepo.full_name,
+        mockIssue.number,
+      );
       expect(res).toEqual(mockIssue);
     });
   });
@@ -614,7 +645,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/labels`)
         .reply(200, [mockLabel, otherMockLabel]);
 
-      const res = await getRepoLabels(mockRepo.full_name);
+      const res = await getRepoLabels(giteaHttp, mockRepo.full_name);
       expect(res).toEqual([mockLabel, otherMockLabel]);
     });
   });
@@ -626,7 +657,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/orgs/${mockRepo.owner.login}/labels`)
         .reply(200, [mockLabel, otherMockLabel]);
 
-      const res = await getOrgLabels(mockRepo.owner.login);
+      const res = await getOrgLabels(giteaHttp, mockRepo.owner.login);
       expect(res).toEqual([mockLabel, otherMockLabel]);
     });
   });
@@ -641,7 +672,12 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200);
 
       await expect(
-        unassignLabel(mockRepo.full_name, mockIssue.number, mockLabel.id),
+        unassignLabel(
+          giteaHttp,
+          mockRepo.full_name,
+          mockIssue.number,
+          mockLabel.id,
+        ),
       ).toResolve();
     });
   });
@@ -656,6 +692,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, mockComment);
 
       const res = await createComment(
+        giteaHttp,
         mockRepo.full_name,
         mockIssue.number,
         mockComment.body,
@@ -677,6 +714,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, updatedMockComment);
 
       const res = await updateComment(
+        giteaHttp,
         mockRepo.full_name,
         mockComment.id,
         'new-body',
@@ -694,7 +732,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         )
         .reply(200);
 
-      const res = await deleteComment(mockRepo.full_name, mockComment.id);
+      const res = await deleteComment(
+        giteaHttp,
+        mockRepo.full_name,
+        mockComment.id,
+      );
       expect(res).toBeUndefined();
     });
   });
@@ -706,7 +748,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/issues/${mockIssue.number}/comments`)
         .reply(200, [mockComment]);
 
-      const res = await getComments(mockRepo.full_name, mockIssue.number);
+      const res = await getComments(
+        giteaHttp,
+        mockRepo.full_name,
+        mockIssue.number,
+      );
       expect(res).toEqual([mockComment]);
     });
   });
@@ -718,12 +764,17 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .post(`/repos/${mockRepo.full_name}/statuses/${mockCommitHash}`)
         .reply(200, mockCommitStatus);
 
-      const res = await createCommitStatus(mockRepo.full_name, mockCommitHash, {
-        state: mockCommitStatus.status,
-        context: mockCommitStatus.context,
-        description: mockCommitStatus.description,
-        target_url: mockCommitStatus.target_url,
-      });
+      const res = await createCommitStatus(
+        giteaHttp,
+        mockRepo.full_name,
+        mockCommitHash,
+        {
+          state: mockCommitStatus.status,
+          context: mockCommitStatus.context,
+          description: mockCommitStatus.description,
+          target_url: mockCommitStatus.target_url,
+        },
+      );
       expect(res).toEqual(mockCommitStatus);
     });
   });
@@ -736,6 +787,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .reply(200, [mockCommitStatus, otherMockCommitStatus]);
 
       const res = await getCombinedCommitStatus(
+        giteaHttp,
         mockRepo.full_name,
         mockBranch.name,
       );
@@ -807,6 +859,7 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         // Expect to get the current state back as the worst status, as all previous commit statuses
         // should be less important than the one which just got added
         const res = await getCombinedCommitStatus(
+          giteaHttp,
           mockRepo.full_name,
           mockBranch.name,
         );
@@ -822,7 +875,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/branches/${mockBranch.name}`)
         .reply(200, mockBranch);
 
-      const res = await getBranch(mockRepo.full_name, mockBranch.name);
+      const res = await getBranch(
+        giteaHttp,
+        mockRepo.full_name,
+        mockBranch.name,
+      );
       expect(res).toEqual(mockBranch);
     });
 
@@ -834,7 +891,11 @@ describe('modules/platform/forgejo/forgejo-helper', () => {
         .get(`/repos/${mockRepo.full_name}/branches/${escapedBranchName}`)
         .reply(200, otherMockBranch);
 
-      const res = await getBranch(mockRepo.full_name, otherMockBranch.name);
+      const res = await getBranch(
+        giteaHttp,
+        mockRepo.full_name,
+        otherMockBranch.name,
+      );
       expect(res).toEqual(otherMockBranch);
     });
   });
