+++ lib/modules/platform/gitea-common/schema.ts
