+++ lib/modules/platform/gitea-common/utils.spec.ts
@@ -1,6 +1,6 @@
 import { partial } from '~test/util.ts';
 import { CONFIG_GIT_URL_UNAVAILABLE } from '../../../constants/error-messages.ts';
-import type { Repo } from './types.ts';
+import type { Repo } from './schema.ts';
 import {
   getMergeMethod,
   getRepoUrl,
@@ -8,11 +8,11 @@ import {
   usableRepo,
 } from './utils.ts';
 
-describe('modules/platform/gitea/utils', () => {
+describe('modules/platform/gitea-common/utils', () => {
   const mockRepo = partial<Repo>({
     allow_rebase: true,
-    clone_url: 'https://gitea.renovatebot.com/some/repo.git',
-    ssh_url: 'git@gitea.renovatebot.com/some/repo.git',
+    clone_url: 'https://forgejo.renovatebot.com/some/repo.git',
+    ssh_url: 'git@forgejo.renovatebot.com/some/repo.git',
     default_branch: 'master',
     full_name: 'some/repo',
     permissions: {
@@ -24,27 +24,27 @@ describe('modules/platform/gitea/utils', () => {
   });
 
   it('trimTrailingApiPath', () => {
-    expect(trimTrailingApiPath('https://gitea.renovatebot.com/api/v1')).toBe(
-      'https://gitea.renovatebot.com/',
+    expect(trimTrailingApiPath('https://forgejo.renovatebot.com/api/v1')).toBe(
+      'https://forgejo.renovatebot.com/',
     );
-    expect(trimTrailingApiPath('https://gitea.renovatebot.com/api/v1/')).toBe(
-      'https://gitea.renovatebot.com/',
+    expect(trimTrailingApiPath('https://forgejo.renovatebot.com/api/v1/')).toBe(
+      'https://forgejo.renovatebot.com/',
     );
-    expect(trimTrailingApiPath('https://gitea.renovatebot.com/')).toBe(
-      'https://gitea.renovatebot.com/',
+    expect(trimTrailingApiPath('https://forgejo.renovatebot.com/')).toBe(
+      'https://forgejo.renovatebot.com/',
     );
-    expect(trimTrailingApiPath('https://gitea.renovatebot.com')).toBe(
-      'https://gitea.renovatebot.com',
+    expect(trimTrailingApiPath('https://forgejo.renovatebot.com')).toBe(
+      'https://forgejo.renovatebot.com',
     );
     expect(
-      trimTrailingApiPath('https://gitea.renovatebot.com/api/gitea/api/v1'),
-    ).toBe('https://gitea.renovatebot.com/api/gitea/');
+      trimTrailingApiPath('https://forgejo.renovatebot.com/api/forgejo/api/v1'),
+    ).toBe('https://forgejo.renovatebot.com/api/forgejo/');
   });
 
   describe('getRepoUrl', () => {
     it('should abort when endpoint is not valid', () => {
       expect.assertions(1);
-      expect(() => getRepoUrl(mockRepo, 'endpoint', 'abc')).toThrow(
+      expect(() => getRepoUrl(mockRepo, 'endpoint', 'abc', 'gitea')).toThrow(
         CONFIG_GIT_URL_UNAVAILABLE,
       );
     });
