+++ lib/workers/repository/update/pr/changelog/forgejo/index.ts
@@ -4,7 +4,7 @@ import { Releases } from '../../../../../../modules/datasource/forgejo-releases/
 import {
   ContentsListResponse,
   RepoContents,
-} from '../../../../../../modules/platform/forgejo/schema.ts';
+} from '../../../../../../modules/platform/gitea-common/schema.ts';
 import { ForgejoHttp } from '../../../../../../util/http/forgejo.ts';
 import { fromBase64 } from '../../../../../../util/string.ts';
 import { compareChangelogFilePath } from '../common.ts';
